users may view/create/update/delete their own tasks

+++ app/main/routes.py
@@ -27,12 +27,14 @@ def get_task(task_id):
 
 
 @bp.route('/todo/api/v1.0/tasks', methods=['POST'])
+@auth_.login_required
 def create_task():
     if not request.json or not 'title' in request.json:
         abort(400)
     
+    user = auth_.current_user()
     task = Task(title=request.json['title'],
-                description=request.json.get('description', ''))
+                description=request.json.get('description', ''), author=user)
     
     db.session.add(task)
     db.session.commit()
@@ -41,10 +43,13 @@ def create_task():
 
 
 @bp.route('/todo/api/v1.0/tasks/<int:task_id>', methods=['PUT'])
+@auth_.login_required
 def update_task(task_id):
     task = Task.query.get(task_id)
     if task is None:
         abort(404)
+    if task.author != auth_.current_user():
+        abort(401)
     if not request.json:
         abort(400)
     if 'title' in request.json and type(request.json['title']) is not str:
@@ -65,10 +70,13 @@ def update_task(task_id):
 
 
 @bp.route('/todo/api/v1.0/tasks/<int:task_id>', methods=['DELETE'])
+@auth_.login_required
 def delete_task(task_id):
     task = Task.query.get(task_id)
     if task is None:
         abort(404)
+    if task.author != auth_.current_user():
+        abort(401)
 
     db.session.delete(task)
     db.session.commit()
